+++ src/Microsoft.Win32.Primitives/src/Interop/Interop.Unix.cs
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+internal partial class Interop
+{
+    internal const string LIBC = "libc";
+}
