+++ src/Microsoft.Win32.Primitives/src/Interop/Interop.Windows.cs
@@ -3,14 +3,13 @@
 
 using System;
 using System.Text;
-using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
 
 internal partial class Interop
 {
     internal partial class mincore
     {
-        [DllImport("api-ms-win-core-localization-l1-2-0.dll", CharSet = CharSet.Unicode, SetLastError = true, BestFitMapping = true)]
+        [DllImport("api-ms-win-core-localization-l1-2-0.dll", CharSet = CharSet.Unicode,  EntryPoint="FormatMessageW", SetLastError = true, BestFitMapping = true)]
         public static extern int FormatMessage(int dwFlags, IntPtr lpSource_mustBeNull, uint dwMessageId,
             int dwLanguageId, StringBuilder lpBuffer, int nSize, IntPtr[] arguments);
 
@@ -19,11 +18,5 @@ internal partial class Interop
         public const int FORMAT_MESSAGE_ARGUMENT_ARRAY = 0x00002000;
 
         public const int ERROR_INSUFFICIENT_BUFFER = 0x7A;
-
-        public const int E_FAIL = unchecked((int)0x80004005);
     }
 }
-
-
-
-
