+++ src/Microsoft.Win32.Primitives/src/Interop/Interop.cs
@@ -1,0 +1,7 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+internal partial class Interop
+{
+    public const int E_FAIL = unchecked((int)0x80004005);
+}
